requirements.txt
PyQt5==5.15.4
pyserial==3.5
numpy==1.21.2
pyautogui==0.9.53
opencv-python==4.5.5.62
pygame==2.6.0
